% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypmh_model_run_mlp.R
\name{cypmh_model_run_mlp}
\alias{cypmh_model_run_mlp}
\title{CYPMH Model - MLP}
\usage{
cypmh_model_run_mlp(cypmh)
}
\arguments{
\item{cypmh}{the cypmh data}
}
\value{
the model
}
\description{
Runs a neural network MLP model on the CYPMH data
}
