% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypmh_model_run_mlp_logistic.R
\name{cypmh_model_run_mlp_logistic}
\alias{cypmh_model_run_mlp_logistic}
\title{CYPMH Model - MLP Logistic}
\usage{
cypmh_model_run_mlp_logistic(cypmh)
}
\arguments{
\item{cypmh}{the cypmh data}
}
\value{
the model
}
\description{
Runs a neural network MLP model on the CYPMH data
}
