% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypmh_model_run_ordinal_regression.R
\name{cypmh_model_run_ordinal_regression}
\alias{cypmh_model_run_ordinal_regression}
\title{CYPMH Model - Ordinal Regression}
\usage{
cypmh_model_run_ordinal_regression(cypmh)
}
\arguments{
\item{cypmh}{the cypmh data}
}
\value{
a list containing the model, the confusion matrix on the test data, and the recipe used
}
\description{
Runs an ordinal regression model on the CYPMH data
}
