% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypmh_model_run_xgboost.R
\name{cypmh_model_run_xgboost}
\alias{cypmh_model_run_xgboost}
\title{CYPMH Model - XGBoost}
\usage{
cypmh_model_run_xgboost(cypmh)
}
\arguments{
\item{cypmh}{the cypmh data}
}
\value{
the model
}
\description{
Runs an xgboost model on the CYPMH data
}
