% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypmh_model_run_xgboost_logistic.R
\name{cypmh_model_run_xgboost_logistic}
\alias{cypmh_model_run_xgboost_logistic}
\title{CYPMH Model - XGBoost Logistic}
\usage{
cypmh_model_run_xgboost_logistic(cypmh)
}
\arguments{
\item{cypmh}{the cypmh data}
}
\value{
the model
}
\description{
Runs an xgboost model on the CYPMH data
}
