% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccg_gss_to_ods.R
\name{get_ccg_gss_to_ods}
\alias{get_ccg_gss_to_ods}
\title{Get CCG GSS to ODS code lookup}
\usage{
get_ccg_gss_to_ods()
}
\value{
tibble
}
\description{
Gets the lookup between GSS (E38) and ODS codes
}
