% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ccg_successors.R
\name{get_ccg_successors}
\alias{get_ccg_successors}
\title{Get CCG Successors}
\usage{
get_ccg_successors()
}
\value{
tibble
}
\description{
Maps CCG's from old codes to new codes where a CCG has been succeeded. Also includes a row for each of the current
CCG's from itself to itself
}
