% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fingertips_data.R
\name{get_fingertips_data}
\alias{get_fingertips_data}
\title{Get Fingertips Data}
\usage{
get_fingertips_data()
}
\value{
tibble
}
\description{
Function to load the required data from the fingertip API.
}
