% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lad_successors.R
\name{get_lad_successors}
\alias{get_lad_successors}
\title{Get LAD Successors}
\usage{
get_lad_successors(stp_lookups)
}
\arguments{
\item{stp_lookups}{the results of the stp_lookups target}
}
\value{
tibble
}
\description{
Maps LAD's from old codes to new codes where a LAD has been succeeded. Also includes a row for each of the current
LAD's from itself to itself. Succession is determined by comparing the 2016 list of LSOA's to LAD's to the 2020
list (\code{stp_lookups}) and finding LAD's that don't apppear in 2020 that appeared in 2016. We then join the two
lists on LSOA. If a LAD is split in two we chose the greatest match, based on which of the "new" LAD's has the
most LSOA's from the "old" LAD.
}
