% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stp_nhsregion_lookup.R
\name{get_stp_nhs_region_lookups}
\alias{get_stp_nhs_region_lookups}
\title{Get STP to NHS region Lookups}
\usage{
get_stp_nhs_region_lookups()
}
\value{
tibble
}
\description{
Get's the STP to NHS region lookups for 2020
}
