% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cypmh.R
\name{process_cypmh}
\alias{process_cypmh}
\title{Process CYPMH}
\usage{
process_cypmh(raw_cypmh, imd, ethnicity, source_referral)
}
\arguments{
\item{raw_cypmh}{results from get_cypmh}

\item{imd}{tibble containing columns lsoa and imd}

\item{ethnicity}{tibble containing columns ethnic_code and ethnicity, to remap}

\item{source_referral}{tibble containing columns source_ref_code and source_referral, to remap}
}
\value{
a tibble
}
\description{
Prepares the raw CYPMH data ready for use
}
