% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypmh_model_run_svm.R
\name{cypmh_model_run_svm}
\alias{cypmh_model_run_svm}
\title{CYPMH Model - SVM}
\usage{
cypmh_model_run_svm(cypmh, kernel)
}
\arguments{
\item{cypmh}{the cypmh data}

\item{kernel}{the kernel type to use with svm}
}
\value{
the model
}
\description{
Runs an svm model on the CYPMH data
}
