% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cypmh.R
\name{get_cypmh}
\alias{get_cypmh}
\title{Get CYPMH}
\usage{
get_cypmh(con_str, table_name)
}
\arguments{
\item{con_str}{the connection string}

\item{table_name}{the table name}
}
\value{
a tibble
}
\description{
Get's the Childrens and Young Persons Mental Health Dataset from the MLCSU datawarehouse
}
